requirements.txt
# Cognify FastAPI Application Dependencies
# Python 3.11+

# FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Environment and configuration
python-dotenv>=1.0.0

# Google Gemini AI SDK
google-genai>=0.2.0

# Data validation
pydantic>=2.0.0

# PDF processing
PyMuPDF>=1.23.0

# CORS support (included with FastAPI but explicit)
# Already handled by fastapi

